+++ airflow/www/templates/airflow/task.html
@@ -5,6 +5,22 @@
     {{ super() }}
     <h4>{{ title }}</h4>
     <div>
+        <h5>Blocking Dependencies</h5>
+        <table class="table table-striped table-bordered">
+            <tr>
+                <th>Dependency</th>
+                <th>Reason</th>
+            </tr>
+            {% for dependency, reason in failed_dep_reasons %}
+                <tr>
+                    <td>{{ dependency }}</td>
+                    <td class='code'>{{ reason }}</td>
+                </tr>
+            {% endfor %}
+        </table>
+        {{ html_code|safe }}
+    </div>
+    <div>
         {% for attr, value in special_attrs_rendered.items() %}
             <h5>Attribute: {{ attr }}</h5>
             {{ value|safe }}
